feat: Enhance GT Material Export Functionality
- Added support for exporting merged GT materials, including OBJ and texture atlases.
- Introduced internal POD structures for raw texture and mesh data.
- Implemented background processing for heavy computations during export.
- Updated UI to reflect export status and prevent multiple concurrent exports.
- Improved logging for better debugging and user feedback.
- Refactored texture sampling and material channel collection methods for clarity and efficiency.
- Added utility functions for writing MTL files and handling texture atlases.

+++ Source/VCCSim/Private/Utils/ImageProcesser.cpp
@@ -156,9 +156,10 @@ void FAsyncNormalEXRSaveTask::DoWork()
 {
     // Use modern UE image API with FImage
     FImage Image;
-    Image.Init(Size.X, Size.Y, ERawImageFormat::RGBA32F);
+    // Assertion failed: Format == ERawImageFormat::RGBA16F if we use RGBA32F and then call AsRGBA16F
+    Image.Init(Size.X, Size.Y, ERawImageFormat::RGBA16F);
 
-    // Convert FLinearColor to FImage data
+    // Convert FFloat16Color to FImage data
     TArrayView64<FFloat16Color> ImageData = Image.AsRGBA16F();
 
     if (ImageData.Num() != NormalPixels.Num())
